+++ src/Vortice.DXGI/Debug/InfoQueueFilterDescription.cs
@@ -5,7 +5,7 @@ using System;
 using System.Runtime.InteropServices;
 using SharpGen.Runtime;
 
-namespace Vortice.DXGI
+namespace Vortice.DXGI.Debug
 {
     public partial class InfoQueueFilterDescription
     {