+++ src/Vortice.DXGI/IDXGISurface2.cs
@@ -8,14 +8,23 @@ namespace Vortice.DXGI
 {
     public partial class IDXGISurface2
     {
-        public T? GetResource<T>(out int subresourceIndex) where T : ComObject
+        public T GetResource<T>(out int subresourceIndex) where T : ComObject
         {
-            if (GetResource(typeof(T).GUID, out IntPtr nativePtr, out subresourceIndex).Failure)
+            GetResource(typeof(T).GUID, out IntPtr nativePtr, out subresourceIndex).CheckError();
+            return MarshallingHelpers.FromPointer<T>(nativePtr);
+        }
+
+        public Result GetResource<T>(out int subresourceIndex, out T? parentResource) where T : ComObject
+        {
+            Result result = GetResource(typeof(T).GUID, out IntPtr nativePtr, out subresourceIndex);
+            if (result.Failure)
             {
-                return default;
+                parentResource = default;
+                return result;
             }
 
-            return MarshallingHelpers.FromPointer<T>(nativePtr);
+            parentResource = MarshallingHelpers.FromPointer<T>(nativePtr);
+            return result;
         }
     }
 }