+++ src/Vortice.Direct3D12/ID3D12Device4.cs
@@ -8,6 +8,7 @@ namespace Vortice.Direct3D12
 {
     public partial class ID3D12Device4
     {
+        #region CreateCommandList1
         public T CreateCommandList1<T>(CommandListType type, CommandListFlags commandListFlags = CommandListFlags.None) where T : ID3D12GraphicsCommandList1
         {
             CreateCommandList1(0, type, commandListFlags, typeof(T).GUID, out IntPtr nativePtr).CheckError();
@@ -37,14 +38,58 @@ namespace Vortice.Direct3D12
             commandList = MarshallingHelpers.FromPointer<T>(nativePtr);
             return result;
         }
+        #endregion
 
-        public T? CreateCommittedResource1<T>(
+        #region CreateCommittedResource1
+        public T CreateCommittedResource1<T>(
             HeapProperties heapProperties,
             HeapFlags heapFlags,
             ResourceDescription description,
             ResourceStates initialResourceState,
             ID3D12ProtectedResourceSession protectedSession,
             ClearValue? optimizedClearValue = null) where T : ID3D12Resource1
+        {
+            CreateCommittedResource1(ref heapProperties, heapFlags,
+                ref description,
+                initialResourceState,
+                optimizedClearValue,
+                protectedSession,
+                typeof(T).GUID, out IntPtr nativePtr).CheckError();
+            return MarshallingHelpers.FromPointer<T>(nativePtr);
+        }
+
+        public Result CreateCommittedResource1<T>(
+            HeapProperties heapProperties,
+            HeapFlags heapFlags,
+            ResourceDescription description,
+            ResourceStates initialResourceState,
+            ID3D12ProtectedResourceSession protectedSession,
+            out T? resource) where T : ID3D12Resource1
+        {
+            Result result = CreateCommittedResource1(ref heapProperties, heapFlags,
+                ref description,
+                initialResourceState,
+                null,
+                protectedSession,
+                typeof(T).GUID, out IntPtr nativePtr);
+            if (result.Failure)
+            {
+                resource = default;
+                return result;
+            }
+
+            resource = MarshallingHelpers.FromPointer<T>(nativePtr);
+            return result;
+        }
+
+        public Result CreateCommittedResource1<T>(
+            HeapProperties heapProperties,
+            HeapFlags heapFlags,
+            ResourceDescription description,
+            ResourceStates initialResourceState,
+            ID3D12ProtectedResourceSession protectedSession,
+            ClearValue optimizedClearValue,
+            out T? resource) where T : ID3D12Resource1
         {
             Result result = CreateCommittedResource1(ref heapProperties, heapFlags,
                 ref description,
@@ -53,45 +98,96 @@ namespace Vortice.Direct3D12
                 protectedSession,
                 typeof(T).GUID, out IntPtr nativePtr);
             if (result.Failure)
-                return default;
+            {
+                resource = default;
+                return result;
+            }
 
+            resource = MarshallingHelpers.FromPointer<T>(nativePtr);
+            return result;
+        }
+        #endregion
+
+        #region CreateHeap1
+        public T CreateHeap1<T>(HeapDescription description, ID3D12ProtectedResourceSession protectedSession) where T : ID3D12Heap1
+        {
+            CreateHeap1(ref description, protectedSession, typeof(T).GUID, out IntPtr nativePtr).CheckError();
             return MarshallingHelpers.FromPointer<T>(nativePtr);
         }
 
-        public T? CreateHeap1<T>(HeapDescription description, ID3D12ProtectedResourceSession protectedSession) where T : ID3D12Heap1
+        public Result CreateHeap1<T>(HeapDescription description, ID3D12ProtectedResourceSession protectedSession, out T? heap) where T : ID3D12Heap1
         {
             Result result = CreateHeap1(ref description, protectedSession, typeof(T).GUID, out IntPtr nativePtr);
             if (result.Failure)
-                return default;
+            {
+                heap = default;
+                return result;
+            }
 
+            heap = MarshallingHelpers.FromPointer<T>(nativePtr);
+            return result;
+        }
+        #endregion
+
+        #region CreateProtectedResourceSession
+        public T CreateProtectedResourceSession<T>(ProtectedResourceSessionDescription description) where T : ID3D12ProtectedResourceSession
+        {
+            CreateProtectedResourceSession(description, typeof(T).GUID, out IntPtr nativePtr).CheckError();
             return MarshallingHelpers.FromPointer<T>(nativePtr);
         }
 
-        public T? CreateProtectedResourceSession<T>(ProtectedResourceSessionDescription description) where T : ID3D12ProtectedResourceSession
+        public Result CreateProtectedResourceSession<T>(ProtectedResourceSessionDescription description, out T? resource) where T : ID3D12ProtectedResourceSession
         {
             Result result = CreateProtectedResourceSession(description, typeof(T).GUID, out IntPtr nativePtr);
             if (result.Failure)
-                return default;
+            {
+                resource = default;
+                return result;
+            }
 
+            resource = MarshallingHelpers.FromPointer<T>(nativePtr);
+            return result;
+        }
+        #endregion
+
+        #region CreateReservedResource1
+        public T CreateReservedResource1<T>(ResourceDescription description, ResourceStates initialState, ClearValue clearValue, ID3D12ProtectedResourceSession protectedResourceSession) where T : ID3D12Resource1
+        {
+            CreateReservedResource1(ref description, initialState, clearValue, protectedResourceSession, typeof(T).GUID, out IntPtr nativePtr).CheckError();
             return MarshallingHelpers.FromPointer<T>(nativePtr);
         }
 
-        public T? CreateReservedResource1<T>(ResourceDescription description, ResourceStates initialState, ClearValue clearValue, ID3D12ProtectedResourceSession protectedResourceSession) where T : ID3D12Resource1
+        public Result CreateReservedResource1<T>(ResourceDescription description, ResourceStates initialState, ClearValue clearValue, ID3D12ProtectedResourceSession protectedResourceSession, out T? resource) where T : ID3D12Resource1
         {
             Result result = CreateReservedResource1(ref description, initialState, clearValue, protectedResourceSession, typeof(T).GUID, out IntPtr nativePtr);
             if (result.Failure)
-                return default;
+            {
+                resource = default;
+                return result;
+            }
 
+            resource = MarshallingHelpers.FromPointer<T>(nativePtr);
+            return result;
+        }
+
+        public T CreateReservedResource1<T>(ResourceDescription description, ResourceStates initialState, ID3D12ProtectedResourceSession protectedResourceSession) where T : ID3D12Resource1
+        {
+            CreateReservedResource1(ref description, initialState, null, protectedResourceSession, typeof(T).GUID, out IntPtr nativePtr).CheckError();
             return MarshallingHelpers.FromPointer<T>(nativePtr);
         }
 
-        public T? CreateReservedResource1<T>(ResourceDescription description, ResourceStates initialState, ID3D12ProtectedResourceSession protectedResourceSession) where T : ID3D12Resource1
+        public Result CreateReservedResource1<T>(ResourceDescription description, ResourceStates initialState, ID3D12ProtectedResourceSession protectedResourceSession, out T? resource) where T : ID3D12Resource1
         {
             Result result = CreateReservedResource1(ref description, initialState, null, protectedResourceSession, typeof(T).GUID, out IntPtr nativePtr);
             if (result.Failure)
-                return default;
+            {
+                resource = default;
+                return result;
+            }
 
-            return MarshallingHelpers.FromPointer<T>(nativePtr);
+            resource = MarshallingHelpers.FromPointer<T>(nativePtr);
+            return result;
         }
+        #endregion
     }
 }