+++ src/Vortice.Direct3D12/ID3D12Device8.cs
@@ -8,29 +8,43 @@ namespace Vortice.Direct3D12
 {
     public partial class ID3D12Device8
     {
-        public T? CreateCommittedResource2<T>(
+        public T CreateCommittedResource2<T>(
             HeapProperties heapProperties,
             HeapFlags heapFlags,
             ResourceDescription1 description,
             ResourceStates initialResourceState,
-            ClearValue? optimizedClearValue,
             ID3D12ProtectedResourceSession protectedSession) where T : ID3D12Resource
         {
-            Result result = CreateCommittedResource2(
+            CreateCommittedResource2(
+                ref heapProperties,
+                heapFlags,
+                ref description,
+                initialResourceState,
+                null,
+                protectedSession,
+                typeof(T).GUID, out IntPtr nativePtr).CheckError();
+
+            return MarshallingHelpers.FromPointer<T>(nativePtr);
+        }
+
+        public T CreateCommittedResource2<T>(
+            HeapProperties heapProperties,
+            HeapFlags heapFlags,
+            ResourceDescription1 description,
+            ResourceStates initialResourceState,
+            ClearValue optimizedClearValue,
+            ID3D12ProtectedResourceSession protectedSession) where T : ID3D12Resource
+        {
+            CreateCommittedResource2(
                 ref heapProperties,
                 heapFlags,
                 ref description,
                 initialResourceState,
                 optimizedClearValue,
                 protectedSession,
-                typeof(T).GUID, out IntPtr nativePtr);
+                typeof(T).GUID, out IntPtr nativePtr).CheckError();
 
-            if (result.Success)
-            {
-                return MarshallingHelpers.FromPointer<T>(nativePtr);
-            }
-
-            return default;
+            return MarshallingHelpers.FromPointer<T>(nativePtr);
         }
 
         public T? CreatePlacedResource1<T>(ID3D12Heap heap, ulong heapOffset, ResourceDescription1 description, ResourceStates initialState, ClearValue? optimizedClearValue = null) where T : ID3D12Resource
@@ -49,6 +63,57 @@ namespace Vortice.Direct3D12
             }
 
             return default;
+        }
+
+        public Result CreatePlacedResource1<T>(
+            ID3D12Heap heap,
+            ulong heapOffset,
+            ResourceDescription1 description,
+            ResourceStates initialState,
+            out T? resource) where T : ID3D12Resource
+        {
+            Result result = CreatePlacedResource1(
+                heap,
+                heapOffset,
+                ref description,
+                initialState,
+                null,
+                typeof(T).GUID, out IntPtr nativePtr);
+
+            if (result.Failure)
+            {
+                resource = default;
+                return result;
+            }
+
+            resource = MarshallingHelpers.FromPointer<T>(nativePtr);
+            return result;
+        }
+
+        public Result CreatePlacedResource1<T>(
+            ID3D12Heap heap,
+            ulong heapOffset,
+            ResourceDescription1 description,
+            ResourceStates initialState,
+            ClearValue optimizedClearValue,
+            out T? resource) where T : ID3D12Resource
+        {
+            Result result = CreatePlacedResource1(
+                heap,
+                heapOffset,
+                ref description,
+                initialState,
+                optimizedClearValue,
+                typeof(T).GUID, out IntPtr nativePtr);
+
+            if (result.Failure)
+            {
+                resource = default;
+                return result;
+            }
+
+            resource = MarshallingHelpers.FromPointer<T>(nativePtr);
+            return result;
         }
     }
 }