+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Current Stable Release:    1.9.80 (May 2021)
 Release:     1.9.XX (June 2021)
 -----------------------------------------------
 - [ENH] Direct3D11: Rename Usage to ResourceUsage and improve CreateBuffer, improve Create shaders with Blob directly.
+- [ENH] DXGI: dxgidebug.h under Vortice.DXGI.Debug namespace.
+- [ENH] Direct3D12: Improve Debug MessageID enum values names.
 
 -----------------------------------------------
 Release:     1.9.80 (May 2021)